requirements.txt
streamlit
streamlit-option-menu
pandas
mysql-connector-python
xlsxwriter
